BackEnd/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.24
Django==4.1.3
djangorestframework==3.14.0
influxdb-client==1.34.0
PyMySQL==1.0.2
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.6
reactivex==4.0.4
six==1.16.0
sqlparse==0.4.3
typing-extensions==4.4.0
urllib3==1.26.13
